refactor(#1373): extract tau-extensions inline tests
